Partially de-duplicate flood_model/*/*_simulations_generator.py
