+++ src/eddie_floodresilience/flood_model/bgflood/bgflood_simulations_generator.py
@@ -23,7 +23,6 @@ import logging
 import platform
 import shutil
 import subprocess
-from datetime import datetime
 from pathlib import Path
 
 import geopandas as gpd
@@ -38,109 +37,16 @@ from src.eddie_floodresilience.flood_model.flooded_buildings import (
     find_flooded_buildings, store_flooded_buildings_in_database
 )
 from src.eddie_floodresilience.flood_model.serve_model import add_model_output_to_geoserver
-from ..flood_model_inputs_generator import InjectionPointsFloodModelGenerator, TerrainGenerator
 from .bgflood_parameters_generator import BGFloodParametersGenerator
 from .bgflood_precipitation import BGFloodPrecipitationGenerator, BGFloodPrecipitationFloodModelGenerator
+from ..flood_model_siumulations_generator import BaseFloodModelSimulationsGenerator
 
 setup_logging(LogLevel.DEBUG)
 log = logging.getLogger(__name__)
 
 
-class BGFloodModelSimulationsGenerator():
+class BGFloodModelSimulationsGenerator(BaseFloodModelSimulationsGenerator):
     """This class is to generate flood model simulations"""  # pylint: disable=too-many-instance-attributes
-
-    def __init__(
-        self,
-        flood_model_path: Path,
-        catchment_model_path: Path,
-        hydromt_path: Path,
-        river_name: str,
-        precipitation_path: Path,
-        aoi_boundary: list,
-        adjust_manning: bool,
-        start_time: datetime,
-        end_time: datetime,
-        crs: int = 2193,
-        polygons: str = None,
-        vectors: str = None
-    ) -> None:
-        """
-        Generate flood model simulations
-
-        Parameters
-        ----------
-        flood_model_path : Path
-            Directory to folder storing terrain data
-        catchment_model_path : Path
-            Directory to folder storing catchment model results
-        hydromt_path : Path
-            A directory to where all necessary files are stored to run wflow model
-        river_name: str
-            Name of directory to where the river information files are stored
-        precipitation_path : Path
-            Directory to folder storing precipitation data
-        aoi_boundary : list
-            Boundaries' coordinates of area of interest.
-            Format is [xmin, ymin, xmax, ymax]
-        adjust_manning : bool
-            True means adjusting Manning's n by resampling 4m Manning's n
-            False means no Mannning's n adjustment
-        start_time: datetime
-            Starting time of the flood event
-        end_time: datetime
-            Ending time of the flood event
-        crs : int = 2193
-            Targeted crs. The default is 2193 for NZTM.
-        polygons : str = None
-            Name of polygon file that is used to change the landcover information.
-            This polygon dataframe has 'landcover' column with new values
-        vectors : str = None
-            Name of vector file that is used to change the elevation information.
-            This vector dataframe has 'value' column to specify increasing or decreasing elevation,
-            and 'distance' column to specify how smooth to decrease elevation.
-        """
-        self.flood_model_path = flood_model_path
-        self.catchment_model_path = catchment_model_path
-        self.hydromt_path = hydromt_path
-        self.river_name = river_name
-        self.precipitation_path = precipitation_path
-        self.aoi_boundary = aoi_boundary
-        self.adjust_manning = adjust_manning
-        self.start_time = start_time
-        self.end_time = end_time
-        self.crs = crs
-        self.polygons = polygons
-        self.vectors = vectors
-
-        # Call out class to generate common terrain data
-        self.terrain = TerrainGenerator(
-            self.flood_model_path,
-            self.hydromt_path,
-            self.river_name,
-            self.aoi_boundary,
-            self.polygons,
-            self.vectors,
-            self.crs
-        )
-
-        # Generate common terrain data
-        self.terrain_bounding_box, self.terrain_crs_clipped = self.terrain.terrain_data_generator()
-
-    def injection_points_for_flood_model_generator(self) -> None:
-        """Generate injection points for flood model (BG_Flood)"""
-        # Call out class used to generate injection points for flood model
-        injection_points_for_flood_model = InjectionPointsFloodModelGenerator(
-            self.flood_model_path,
-            self.catchment_model_path,
-            self.terrain_bounding_box,
-            self.start_time,
-            self.end_time,
-            self.polygons,
-            self.crs
-        )
-
-        # Generate injection points for flood model
-        injection_points_for_flood_model.injection_points_flow_generator()
 
     def precipitation_data_for_flood_model_generator(self) -> None:
         """Generate precipitation data for flood model"""
@@ -175,7 +81,7 @@ class BGFloodModelSimulationsGenerator():
             # Generate precipitation data for flood model
             precipitation_data_for_flood_model.precipitation_for_flood_model_generator()
 
-    def parameter_data_for_flood_model_generator(self) -> None:
+    def parameter_files_for_flood_model_generator(self) -> None:
         """Generate parameters files for flood model"""
         # Call out class used to generate parameter files
         parameters_files_generator = BGFloodParametersGenerator(
@@ -224,14 +130,19 @@ class BGFloodModelSimulationsGenerator():
 
         return output_folder_path
 
-    def flood_model_simulations_generator(self) -> int:
+    def flood_model_simulations_generator(self, output_dir: Path | None) -> int:
         """
         Generate flood simulations by running flood model
+
+        Parameters
+        ----------
+        output_dir : Path
+            The path to the output directory, to allow for serving.
 
         Returns
         -------
         int
-            The model output ID of the flood model run
+            The Flood Model output ID
         """
         # Set up path to log file
         log_file_path = self.flood_model_path / "simulation_log.log"
@@ -332,13 +243,13 @@ class BGFloodModelSimulationsGenerator():
             self.injection_points_for_flood_model_generator()
 
             # Generate parameter files for flood model
-            self.parameter_data_for_flood_model_generator()
+            self.parameter_files_for_flood_model_generator()
 
             # Generate precipitation data for flood model
             self.precipitation_data_for_flood_model_generator()
 
             # Generate simulations by running flood model
-            model_output_id = self.flood_model_simulations_generator()
+            model_output_id = self.flood_model_simulations_generator(None)
 
         # This 'elif' includes 3 and 4
         elif self.polygons is not None or self.vectors is None:
@@ -346,16 +257,16 @@ class BGFloodModelSimulationsGenerator():
             self.injection_points_for_flood_model_generator()
 
             # Generate parameter files for flood model
-            self.parameter_data_for_flood_model_generator()
+            self.parameter_files_for_flood_model_generator()
 
             # Generate simulations by running flood model
-            model_output_id = self.flood_model_simulations_generator()
+            model_output_id = self.flood_model_simulations_generator(None)
 
         # This 'else' includes 2
         else:
             # Generate parameter files for flood model
-            self.parameter_data_for_flood_model_generator()
+            self.parameter_files_for_flood_model_generator()
 
             # Generate simulations by running flood model
-            model_output_id = self.flood_model_simulations_generator()
+            model_output_id = self.flood_model_simulations_generator(None)
         return model_output_id
